+++ src/pluto_main/Table_CannedEvents_CriteriaParmList.cpp
@@ -369,7 +369,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(100,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)100,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -383,7 +383,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comments.c_str(), (unsigned long) min(100,m_Comments.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comments.c_str(), (unsigned long) min((size_t)100,m_Comments.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -410,7 +410,7 @@ if (is_null[6])
 return "NULL";
 
 char *buf = new char[41];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Parm.c_str(), (unsigned long) min(20,m_Parm.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Parm.c_str(), (unsigned long) min((size_t)20,m_Parm.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -424,7 +424,7 @@ if (is_null[7])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_DefaultValue.c_str(), (unsigned long) min(50,m_DefaultValue.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_DefaultValue.c_str(), (unsigned long) min((size_t)50,m_DefaultValue.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -438,7 +438,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ExtraInfo.c_str(), (unsigned long) min(65535,m_ExtraInfo.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ExtraInfo.c_str(), (unsigned long) min((size_t)65535,m_ExtraInfo.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -517,7 +517,7 @@ if (is_null[14])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
