+++ src/pluto_main/Table_CommandGroup.cpp
@@ -478,7 +478,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(50,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)50,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -492,7 +492,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Hint.c_str(), (unsigned long) min(30,m_Hint.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Hint.c_str(), (unsigned long) min((size_t)30,m_Hint.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -714,7 +714,7 @@ if (is_null[21])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
