+++ src/pluto_main/Table_CriteriaParm.cpp
@@ -334,7 +334,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[41];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Parm.c_str(), (unsigned long) min(20,m_Parm.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Parm.c_str(), (unsigned long) min((size_t)20,m_Parm.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -361,7 +361,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Value.c_str(), (unsigned long) min(50,m_Value.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Value.c_str(), (unsigned long) min((size_t)50,m_Value.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -440,7 +440,7 @@ if (is_null[11])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
