+++ src/pluto_main/Table_Device.cpp
@@ -548,7 +548,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(30,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)30,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -588,7 +588,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[31];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_IPaddress.c_str(), (unsigned long) min(15,m_IPaddress.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_IPaddress.c_str(), (unsigned long) min((size_t)15,m_IPaddress.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -602,7 +602,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[37];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_MACaddress.c_str(), (unsigned long) min(18,m_MACaddress.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_MACaddress.c_str(), (unsigned long) min((size_t)18,m_MACaddress.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -655,7 +655,7 @@ if (is_null[13])
 return "NULL";
 
 char *buf = new char[81];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_State.c_str(), (unsigned long) min(40,m_State.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_State.c_str(), (unsigned long) min((size_t)40,m_State.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -669,7 +669,7 @@ if (is_null[14])
 return "NULL";
 
 char *buf = new char[81];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Status.c_str(), (unsigned long) min(40,m_Status.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Status.c_str(), (unsigned long) min((size_t)40,m_Status.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -774,7 +774,7 @@ if (is_null[22])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
