+++ src/pluto_main/Table_Device_QuickStart.cpp
@@ -354,7 +354,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(50,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)50,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -381,7 +381,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Binary.c_str(), (unsigned long) min(50,m_Binary.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Binary.c_str(), (unsigned long) min((size_t)50,m_Binary.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -395,7 +395,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Arguments.c_str(), (unsigned long) min(255,m_Arguments.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Arguments.c_str(), (unsigned long) min((size_t)255,m_Arguments.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -487,7 +487,7 @@ if (is_null[12])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
