+++ src/pluto_main/Table_Document_Comment.cpp
@@ -304,7 +304,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[5000000];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comment.c_str(), (unsigned long) min(16777215,m_Comment.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comment.c_str(), (unsigned long) min((size_t)16777215,m_Comment.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -318,7 +318,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[25];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Date.c_str(), (unsigned long) min(12,m_Date.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Date.c_str(), (unsigned long) min((size_t)12,m_Date.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -384,7 +384,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
